backend/services/video_generator.py
import os
import time
import asyncio
from typing import Optional, Dict, Any
import google.generativeai as genai
from config.settings import settings


class VideoGenerator:
    def __init__(self):
        # Configure Google Generative AI
        genai.configure(api_key=settings.google_api_key)
        self.model = settings.veo_model
    
    async def generate_video(self, prompt: str, negative_prompt: Optional[str] = None) -> Dict[str, Any]:
        """
        Generate a video using Google Veo API
        """
        try:
            # Create uploads directory if it doesn't exist
            os.makedirs(settings.upload_dir, exist_ok=True)
            
            # Generate video using Google Veo
            # Note: This is a placeholder for the actual Veo API integration
            # The current google-generativeai library doesn't support Veo yet
            # When Veo API is available, this would use the actual API
            
            # For now, we'll create a placeholder video file
            timestamp = int(time.time())
            video_path = os.path.join(settings.upload_dir, f"generated_video_{timestamp}.mp4")
            
            # Create a placeholder file (in production, this would be the actual generated video)
            with open(video_path, 'w') as f:
                f.write(f"Placeholder video for prompt: {prompt}")
            
            return {
                "success": True,
                "video_path": video_path,
                "prompt": prompt,
                "model": self.model
            }
            
        except Exception as e:
            return {
                "success": False,
                "error": str(e),
                "prompt": prompt
            }
    

    
    def validate_prompt(self, prompt: str) -> Dict[str, Any]:
        """
        Validate prompt before sending to Veo API
        """
        if len(prompt) > 1024:
            return {
                "valid": False,
                "error": "Prompt exceeds 1024 token limit"
            }
        
        if not prompt.strip():
            return {
                "valid": False,
                "error": "Prompt cannot be empty"
            }
        
        # Check for safety filter violations
        unsafe_keywords = ["violence", "explicit", "inappropriate"]
        if any(keyword in prompt.lower() for keyword in unsafe_keywords):
            return {
                "valid": False,
                "error": "Prompt contains potentially unsafe content"
            }
        
        return {"valid": True}
