backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.8.0
twelvelabs==0.4.0
aiofiles==23.2.1
pydantic>=2.6.0
httpx==0.25.2
asyncio==3.4.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy>=2.0.25
alembic>=1.13.1
